Fix fee and SPL token amount conversion from Helius API
- Convert fee from lamports to SOL (divide by 10^9)
- Convert SOL amount from lamports to SOL
- Fix SPL tokenAmount precision (use BigDecimal instead of toLong)
- Convert human-readable tokenAmount to raw format for compatibility
- Update Kotlin to 1.9.22 and Room to 2.6.1 for compatibility

+++ solanakit/src/main/java/io/horizontalsystems/solanakit/transactions/HeliusClient.kt
@@ -326,7 +326,7 @@ data class HeliusTransaction(
                 fee = fee.toString(),
                 tokenAccountAddress = transfer.toTokenAccount.ifEmpty { transfer.fromTokenAccount },
                 mintAccountAddress = transfer.mint,
-                splBalanceChange = transfer.tokenAmount.toLong().toString(),
+                splBalanceChange = transfer.tokenAmount.toString(),
                 error = transactionError,
                 splTransferSource = transfer.fromUserAccount,
                 splTransferDestination = transfer.toUserAccount
